+++ Classes/ViewHelpers/SalutationAwareTranslateViewHelper.php
@@ -5,9 +5,7 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\ViewHelpers;
 
 use TYPO3\CMS\Extbase\Utility\LocalizationUtility;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
 use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
 
 /**
  * This class works like the `translate` view helper from Fluid with these two differences:
@@ -24,8 +22,6 @@ use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithRenderStatic;
  */
 class SalutationAwareTranslateViewHelper extends AbstractViewHelper
 {
-    use CompileWithRenderStatic;
-
     private const EXTENSION_NAME = 'seminars';
 
     private const DEFAULT_SALUTATION = 'formal';
@@ -43,46 +39,33 @@ class SalutationAwareTranslateViewHelper extends AbstractViewHelper
         $this->registerArgument('arguments', 'array', 'Arguments to be replaced in the resulting string');
     }
 
-    /**
-     * @param array<string, mixed> $arguments
-     */
-    public static function renderStatic(
-        array $arguments,
-        \Closure $renderChildrenClosure,
-        RenderingContextInterface $renderingContext
-    ): string {
-        $key = self::retrieveKeyFromArguments($arguments);
-        $keyWithSalutation = self::buildKeyWithSalutation($arguments, $renderingContext);
-        $translateArguments = \is_array($arguments['arguments']) ? $arguments['arguments'] : [];
-        $labelWithSalutation = self::translate($keyWithSalutation, $translateArguments);
+    public function render(): string
+    {
+        $key = $this->retrieveKeyFromArguments();
+        $keyWithSalutation = $this->buildKeyWithSalutation();
+        $translateArguments = \is_array($this->arguments['arguments']) ? $this->arguments['arguments'] : [];
+        $labelWithSalutation = $this->translate($keyWithSalutation, $translateArguments);
 
         return ($labelWithSalutation !== $keyWithSalutation)
             ? $labelWithSalutation
-            : self::translate($key, $translateArguments);
+            : $this->translate($key, $translateArguments);
+    }
+
+    private function buildKeyWithSalutation(): string
+    {
+        $salutation = $this->getSalutationMode();
+
+        return $this->retrieveKeyFromArguments() . '_' . $salutation;
     }
 
     /**
-     * @param array<string, mixed> $arguments
-     */
-    private static function buildKeyWithSalutation(
-        array $arguments,
-        RenderingContextInterface $renderingContext
-    ): string {
-        $salutation = self::getSalutationMode($renderingContext);
-
-        return self::retrieveKeyFromArguments($arguments) . '_' . $salutation;
-    }
-
-    /**
-     * @param array<string, mixed> $arguments
-     *
      * @return non-empty-string
      */
-    private static function retrieveKeyFromArguments(array $arguments): string
+    private function retrieveKeyFromArguments(): string
     {
-        \assert(isset($arguments['key']));
+        \assert(isset($this->arguments['key']));
 
-        $result = $arguments['key'];
+        $result = $this->arguments['key'];
         \assert(\is_string($result) && $result !== '');
 
         return $result;
@@ -91,9 +74,9 @@ class SalutationAwareTranslateViewHelper extends AbstractViewHelper
     /**
      * @return non-empty-string
      */
-    private static function getSalutationMode(RenderingContextInterface $renderingContext): string
+    private function getSalutationMode(): string
     {
-        $settings = $renderingContext->getVariableProvider()->get('settings');
+        $settings = $this->renderingContext->getVariableProvider()->get('settings');
 
         return (is_array($settings) && isset($settings['salutation']) && is_string($settings['salutation'])
             && $settings['salutation'] !== '')
@@ -104,7 +87,7 @@ class SalutationAwareTranslateViewHelper extends AbstractViewHelper
     /**
      * @param array<mixed> $arguments
      */
-    public static function translate(string $key, array $arguments): string
+    public function translate(string $key, array $arguments): string
     {
         $value = LocalizationUtility::translate($key, self::EXTENSION_NAME, $arguments);
         if (!is_string($value)) {